+++ ft-sdk/src/lib.rs
@@ -30,7 +30,7 @@ pub use email::{send_email, EmailError};
 pub use error::{not_found_, server_error_, single_error, unauthorised_, SpecialError};
 #[cfg(feature = "field-extractors")]
 pub use from_request::{AppUrl, Cookie, Hidden, Optional, Query, Required};
-pub use from_request::{Form, FromRequest, Host, Path, WrappedFromRequest};
+pub use from_request::{Form, FromRequest, Host, Path, Scheme, WrappedFromRequest};
 pub use ft_derive::{data, form, processor, wrapped_processor};
 #[cfg(feature = "postgres")]
 pub use ft_sys::PgConnection;